Finish C part

+++ main/include/utils.h
@@ -17,8 +17,8 @@
 
 struct CodePos
 {
-	int code;
-	int pos;
+	uint64_t code;
+	uint64_t pos;
 };
 
 class GInfo
